Conflicts:
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/ModuleDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/OptionDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/ProductDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/ProductTagDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/RoleDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/RoleOptionDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/TagDTO.java
	salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/UserDTO.java

+++ salak.business/src/main/java/com/bambucodev/salak/model/Module.java
@@ -26,7 +26,7 @@ public class Module implements Serializable {
 	@SequenceGenerator(name="MODULE_IDMODULE_GENERATOR", sequenceName="SE_MODULE")
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="MODULE_IDMODULE_GENERATOR")
 	@Column(name="id_module")
-	private Integer idModule;
+	private Long idModule;
 
 	private String name;
 
@@ -39,11 +39,11 @@ public class Module implements Serializable {
 	public Module() {
 	}
 
-	public Integer getIdModule() {
+	public Long getIdModule() {
 		return this.idModule;
 	}
 
-	public void setIdModule(Integer idModule) {
+	public void setIdModule(Long idModule) {
 		this.idModule = idModule;
 	}
