+++ salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/ModuleDTO.java
@@ -9,11 +9,9 @@ import java.util.List;
  */
 public class ModuleDTO implements java.io.Serializable {
 
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = -1851486913078641432L;
-	private int idModule;
+
+	private Long idModule;
 	private String name;
 	private boolean state;
 	private List<OptionDTO> options = new ArrayList<OptionDTO>(0);
@@ -21,24 +19,24 @@ public class ModuleDTO implements java.io.Serializable {
 	public ModuleDTO() {
 	}
 
-	public ModuleDTO(int idModule, String name, boolean state) {
+	public ModuleDTO(Long idModule, String name, boolean state) {
 		this.idModule = idModule;
 		this.name = name;
 		this.state = state;
 	}
 
-	public ModuleDTO(int idModule, String name, boolean state, List<OptionDTO> options) {
+	public ModuleDTO(Long idModule, String name, boolean state, List<OptionDTO> options) {
 		this.idModule = idModule;
 		this.name = name;
 		this.state = state;
 		this.options = options;
 	}
 
-	public int getIdModule() {
+	public Long getIdModule() {
 		return this.idModule;
 	}
 
-	public void setIdModule(int idModule) {
+	public void setIdModule(Long idModule) {
 		this.idModule = idModule;
 	}
