+++ salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/ProductDTO.java
@@ -9,11 +9,10 @@ import java.util.List;
  */
 public class ProductDTO implements java.io.Serializable {
 
-	/**
-	 * 
-	 */
+
 	private static final long serialVersionUID = 4093915966274600969L;
-	private int idProduct;
+	
+	private Long idProduct;
 	private CategoryDTO category;
 	private String name;
 	private String description;
@@ -24,7 +23,7 @@ public class ProductDTO implements java.io.Serializable {
 	public ProductDTO() {
 	}
 
-	public ProductDTO(int idProduct, CategoryDTO category, String name, String description, boolean state) {
+	public ProductDTO(Long idProduct, CategoryDTO category, String name, String description, boolean state) {
 		this.idProduct = idProduct;
 		this.category = category;
 		this.name = name;
@@ -32,7 +31,7 @@ public class ProductDTO implements java.io.Serializable {
 		this.state = state;
 	}
 
-	public ProductDTO(int idProduct, CategoryDTO category, String name, String description, Integer price, boolean state,
+	public ProductDTO(Long idProduct, CategoryDTO category, String name, String description, Integer price, boolean state,
 			List<ProductTagDTO> productTags) {
 		this.idProduct = idProduct;
 		this.category = category;
@@ -43,11 +42,11 @@ public class ProductDTO implements java.io.Serializable {
 		this.productTags = productTags;
 	}
 
-	public int getIdProduct() {
+	public Long getIdProduct() {
 		return this.idProduct;
 	}
 
-	public void setIdProduct(int idProduct) {
+	public void setIdProduct(Long idProduct) {
 		this.idProduct = idProduct;
 	}
