+++ salak.illusion/src/main/java/com/bambucodev/salak/illusion/dto/RoleDTO.java
@@ -9,11 +9,10 @@ import java.util.List;
  */
 public class RoleDTO implements java.io.Serializable {
 
-	/**
-	 * 
-	 */
+
 	private static final long serialVersionUID = -6789268300850921114L;
-	private int idRole;
+	
+	private Long idRole;
 	private String name;
 	private boolean state;
 	private String description;
@@ -23,13 +22,13 @@ public class RoleDTO implements java.io.Serializable {
 	public RoleDTO() {
 	}
 
-	public RoleDTO(int idRole, String name, boolean state) {
+	public RoleDTO(Long idRole, String name, boolean state) {
 		this.idRole = idRole;
 		this.name = name;
 		this.state = state;
 	}
 
-	public RoleDTO(int idRole, String name, boolean state, String description, RoleOptionDTO roleOption, List<UserDTO> users) {
+	public RoleDTO(Long idRole, String name, boolean state, String description, RoleOptionDTO roleOption, List<UserDTO> users) {
 		this.idRole = idRole;
 		this.name = name;
 		this.state = state;
@@ -38,11 +37,11 @@ public class RoleDTO implements java.io.Serializable {
 		this.users = users;
 	}
 
-	public int getIdRole() {
+	public Long getIdRole() {
 		return this.idRole;
 	}
 
-	public void setIdRole(int idRole) {
+	public void setIdRole(Long idRole) {
 		this.idRole = idRole;
 	}
